+++ libs/core/src/handlers/events.ts
@@ -43,6 +43,7 @@ export enum SeelenEvent {
   StateWidgetsChanged = "widgets-changed",
   StateWallpapersChanged = "UserResources::wallpapers-changed",
   StateProfilesChanged = "profiles-changed",
+  SystemTrayChanged = "system-tray::changed",
   StatePerformanceModeChanged = "state::performance-mode-changed",
   WidgetTriggered = "widget::triggered",
 }