+++ libs/ui/svelte/components/Icon/common.ts
@@ -1,0 +1,3 @@
+import { IconPackManager } from "@seelen-ui/lib";
+
+export const iconPackManager = await IconPackManager.create();