+++ libs/ui/svelte/components/Icon/index.ts
@@ -1,2 +1,6 @@
 export { default as Icon } from "./Icon.svelte";
 export { default as InlineSVG } from "./InlineSVG.svelte";
+export { default as MissingIcon } from "./MissingIcon.svelte";
+export { default as SpecificIcon } from "./SpecificIcon.svelte";
+export { default as FileIcon } from "./FileIcon.svelte";
+export { iconPackManager } from "./common";