+++ src/ui/react/wallpaper_manager/index.tsx
@@ -1,7 +1,6 @@
 import { startThemingTool } from "@seelen-ui/lib";
 import { getRootContainer } from "@shared";
 import { disableAnimationsOnPerformanceMode } from "@shared/performance";
-import { removeDefaultWebviewActions } from "@shared/setup";
 import { createRoot } from "react-dom/client";
 
 import { App } from "./app.tsx";
@@ -10,7 +9,6 @@ import "@shared/styles/colors.css";
 import "@shared/styles/reset.css";
 import "./styles/global.css";
 
-removeDefaultWebviewActions();
 await startThemingTool();
 disableAnimationsOnPerformanceMode();