+++ src/ui/svelte/system-tray/state.svelte.ts
@@ -1,0 +1,9 @@
+import { invoke, SeelenCommand, SeelenEvent, subscribe } from "@seelen-ui/lib";
+
+export const state = $state({
+  trayItems: await invoke(SeelenCommand.GetSystemTrayIcons),
+});
+
+subscribe(SeelenEvent.SystemTrayChanged, (e) => {
+  state.trayItems = e.payload;
+});